refactor: :recycle: bundle derecursify and remove need for publishing
