+++ packages/delivery-react-native/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@bugsnag/delivery-react-native",
   "version": "8.4.0",
-  "main": "delivery.js",
+  "main": "dist/delivery.js",
   "description": "@bugsnag/js delivery mechanism for React Native",
   "homepage": "https://www.bugsnag.com/",
   "repository": {
@@ -12,8 +12,11 @@
     "access": "public"
   },
   "files": [
-    "*.js"
+    "dist"
   ],
+  "scripts": {
+    "build": "rollup --config rollup.config.js"
+  },
   "author": "Bugsnag",
   "license": "MIT",
   "devDependencies": {
@@ -21,7 +24,6 @@
     "@bugsnag/derecursify": "^8.4.0"
   },
   "peerDependencies": {
-    "@bugsnag/core": "^8.0.0",
-    "@bugsnag/derecursify": "^8.4.0"
+    "@bugsnag/core": "^8.0.0"
   }
 }
