+++ packages/plugin-react-native-client-sync/src/client-sync.js
